+++ src/main/java/uk/gov/pay/api/service/PaymentSearchService.java
@@ -1,18 +1,34 @@
 package uk.gov.pay.api.service;
 
+import black.door.hate.HalRepresentation;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import uk.gov.pay.api.auth.Account;
-import uk.gov.pay.api.model.search.card.SearchCardPayments;
+import uk.gov.pay.api.exception.SearchPaymentsException;
+import uk.gov.pay.api.model.search.PaginationDecorator;
+import uk.gov.pay.api.model.search.card.PaymentForSearchResult;
+import uk.gov.pay.api.model.search.card.PaymentSearchResponse;
 
 import javax.inject.Inject;
+import javax.ws.rs.ProcessingException;
+import javax.ws.rs.client.Client;
+import javax.ws.rs.core.HttpHeaders;
 import javax.ws.rs.core.Response;
 import java.util.LinkedHashMap;
+import java.util.List;
 import java.util.Map;
+import java.util.stream.Collectors;
 
+import static javax.ws.rs.core.MediaType.APPLICATION_JSON;
 import static org.apache.commons.lang3.StringUtils.isNotBlank;
+import static org.apache.http.HttpStatus.SC_OK;
 import static uk.gov.pay.api.validation.PaymentSearchValidator.validateSearchParameters;
 
 public class PaymentSearchService {
 
+    private static final String PAYMENTS_PATH = "/v1/payments";
+    private static final Logger LOGGER = LoggerFactory.getLogger(PaymentSearchService.class);
+    
     public static final String REFERENCE_KEY = "reference";
     public static final String EMAIL_KEY = "email";
     public static final String STATE_KEY = "state";
@@ -25,11 +41,20 @@ public class PaymentSearchService {
     public static final String PAGE = "page";
     public static final String DISPLAY_SIZE = "display_size";
     
-    private final SearchCardPayments searchCardPayments;
+    private final ConnectorUriGenerator connectorUriGenerator;
+    private final Client client;
+    private final PublicApiUriGenerator publicApiUriGenerator;
+    private final PaginationDecorator paginationDecorator;
 
     @Inject
-    public PaymentSearchService(SearchCardPayments searchCardPayments) {
-        this.searchCardPayments = searchCardPayments;
+    public PaymentSearchService(Client client,
+                                ConnectorUriGenerator connectorUriGenerator,
+                                PublicApiUriGenerator publicApiUriGenerator,
+                                PaginationDecorator paginationDecorator) {
+        this.client = client;
+        this.connectorUriGenerator = connectorUriGenerator;
+        this.publicApiUriGenerator = publicApiUriGenerator;
+        this.paginationDecorator = paginationDecorator;
     }
     
     public Response doSearch(Account account, String reference, String email, String state, String cardBrand,
@@ -54,6 +79,46 @@ public class PaymentSearchService {
         queryParams.put(PAGE, pageNumber);
         queryParams.put(DISPLAY_SIZE, displaySize);
         
-        return searchCardPayments.getSearchResponse(account, queryParams);
+        return getSearchResponse(account, queryParams);
+    }
+    
+    private Response getSearchResponse(Account account, Map<String, String> queryParams) {
+        queryParams.put("transactionType", "charge");
+
+        String url = connectorUriGenerator.chargesURIWithParams(account, queryParams);
+        Response connectorResponse = client
+                .target(url)
+                .request()
+                .header(HttpHeaders.ACCEPT, APPLICATION_JSON)
+                .get();
+        LOGGER.info("response from connector for transaction search: {}", connectorResponse);
+        if (connectorResponse.getStatus() == SC_OK) {
+            return processResponse(connectorResponse);
+        }
+        throw new SearchPaymentsException(connectorResponse);
+    }
+
+    private Response processResponse(Response connectorResponse) {
+        PaymentSearchResponse response;
+        try {
+            response = connectorResponse.readEntity(PaymentSearchResponse.class);
+        } catch (ProcessingException ex) {
+            throw new SearchPaymentsException(ex);
+        }
+        List<PaymentForSearchResult> chargeFromResponses = response.getPayments()
+                .stream()
+                .map(charge -> PaymentForSearchResult.valueOf(
+                        charge,
+                        publicApiUriGenerator.getPaymentURI(charge.getChargeId()),
+                        publicApiUriGenerator.getPaymentEventsURI(charge.getChargeId()),
+                        publicApiUriGenerator.getPaymentCancelURI(charge.getChargeId()),
+                        publicApiUriGenerator.getPaymentRefundsURI(charge.getChargeId()),
+                        publicApiUriGenerator.getPaymentCaptureURI(charge.getChargeId())))
+                .collect(Collectors.toList());
+        HalRepresentation.HalRepresentationBuilder halRepresentation = HalRepresentation
+                .builder()
+                .addProperty("results", chargeFromResponses);
+
+        return Response.ok().entity(paginationDecorator.decoratePagination(halRepresentation, response, PAYMENTS_PATH).build().toString()).build();
     }
 }
